fix: remove short

+++ get-array-values.c
@@ -2,18 +2,18 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void print_array(short int *array, unsigned short int size) {
+void print_array(int *array, unsigned short int size) {
   size_t i;
   for (i = 0; i < size; i++) {
     printf("%d ", array[i]);
   }
 }
 
-short int *get_array_values(size_t size) {
+int *get_array_values(size_t size) {
   size_t i;
-  short int *input_array = (short int *)calloc(size, sizeof(short int));
+  int *input_array = (int *)calloc(size, sizeof(int));
   for (i = 0; i < size; i++) {
-    scanf("%hd", &input_array[i]);
+    scanf("%d", &input_array[i]);
   }
   printf("The input array is: ");
   print_array(input_array, size);
