+++ get-array-values.h
@@ -1,3 +1,3 @@
 #include <stddef.h>
 
-short int *get_array_values(size_t size);
+int *get_array_values(size_t size);
