+++ partial-sums.c
@@ -4,7 +4,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int *partial_sums(short int *input_array, unsigned short int array_size) {
+int *partial_sums(int *input_array, unsigned short int array_size) {
   int *accumulated_array;
 
   return accumulated_array;
@@ -12,7 +12,7 @@ int *partial_sums(short int *input_array, unsigned short int array_size) {
 
 int main() {
   size_t arr_len;
-  short int *input_array;
+  int *input_array;
 
   arr_len = get_array_size();
   input_array = get_array_values(arr_len);
